Crate massa_graph

graph management

Modules

error

graph errors

export_active_block

useful structures

settings

graph settings

Structs

BlockGraph

Graph management

BlockGraphExport

Bootstrap compatible version of the block graph

BootstrapableGraph

Bootstrap graph

BootstrapableGraphDeserializer

Basic deserializer for BootstrapableGraph

BootstrapableGraphSerializer

Basic serializer for BootstrapableGraph

ExportCompiledBlock

The block version that can be exported. Note that the detailed list of operation is not exported

Enums

DiscardReason

Something can be discarded

ExportBlockStatus

Block status in the graph that can be exported.

Status

Status

Functions

create_genesis_block

Creates genesis block in given thread.