Struct massa_graph::ledger::Ledger

pub struct Ledger { /* private fields */ }

Here we map an address to its balance. When a balance becomes final it is written on the disk.

Implementations

impl Ledger

pub fn new(
cfg: LedgerConfig,
opt_init_data: Option<LedgerSubset>
) -> Result<Ledger>

if no latest_final_periods in file, they are initialized at 0u64 if there is a ledger in the given file, it is loaded

pub fn get_final_data(&self, addresses: Set<Address>) -> Result<LedgerSubset>

Returns the final ledger data of a list of unique addresses belonging to any thread.

pub fn from_export(
export: LedgerSubset,
latest_final_periods: Vec<u64>,
cfg: LedgerConfig
) -> Result<Ledger>

If there is something in the ledger file, it is overwritten

pub fn get_final_balance(&self, address: &Address) -> Result<Amount>

Returns the final balance of an address. 0 if the address does not exist.

pub fn apply_final_changes(
&self,
thread: u8,
changes: &LedgerChanges,
latest_final_period: u64
) -> Result<()>

Atomically apply a batch of changes to the ledger. All changes should occure in one thread. Update last final period.

If the balance of an address falls exactly to 0, it is removed from the ledger.

If the balance of a non-existing address increases, the address is added to the ledger.

If we attempt to substract more than the balance of an address, the transaction is cancelled and the function returns an error.

pub fn get_latest_final_periods(&self) -> Result<Vec<u64>>

returns the final periods.

pub fn clear(&self) -> Result<()>

To empty the db.

pub fn read_whole(&self) -> Result<LedgerSubset>

Used for bootstrap. Note: this cannot be done transactionally.

pub fn get_final_ledger_subset(
&self,
query_addrs: &Set<Address>
) -> Result<LedgerSubset>

Gets ledger at latest final blocks for query_addrs