Struct massa_graph::BlockGraph

pub struct BlockGraph { /* private fields */ }

Graph management

Implementations

impl BlockGraph

pub async fn new(
cfg: GraphConfig,
init: Option<BootstrapableGraph>,
storage: Storage,
selector_controller: Box<dyn SelectorController>
) -> Result<Self>

Creates a new BlockGraph.

Argument

cfg: consensus configuration.

init: A bootstrap graph to start the graph with

storage: A shared storage that share data across all modules.

selector_controller: Access to the PoS selector to get draws

pub fn export_bootstrap_graph(&self) -> Result<BootstrapableGraph>

export full graph in a bootstrap compatible version

pub fn get_latest_final_blocks_periods(&self) -> &Vec<(BlockId, u64)>

Gets latest final blocks (hash, period) for each thread.

pub fn get_best_parents(&self) -> &Vec<(BlockId, u64)>

Gets best parents.

pub fn get_cliques(&self) -> Vec<Clique>

Gets the list of cliques.

pub fn get_block_ids_by_creator(
&self,
address: &Address
) -> PreHashMap<BlockId, Status>

Returns the list of block IDs created by a given address, and their finality statuses

pub fn get_active_block(
&self,
block_id: &BlockId
) -> Option<(&ActiveBlock, &Storage)>

Gets whole compiled block corresponding to given hash, if it is active.

Argument

block_id: block ID

pub fn get_block_status(&self, block_id: &BlockId) -> BlockGraphStatus

get block graph status

pub fn slot_tick(&mut self, current_slot: Option<Slot>) -> Result<()>

signal new slot

pub fn incoming_header(
&mut self,
block_id: BlockId,
header: WrappedHeader,
current_slot: Option<Slot>
) -> Result<()>

A new header has come !

Checks performed:

Ignore genesis blocks.

See process.

pub fn incoming_block(
&mut self,
block_id: BlockId,
slot: Slot,
current_slot: Option<Slot>,
storage: Storage
) -> Result<()>

A new block has come

Checks performed:

Ignore genesis blocks.

See process.

pub fn invalid_block(
&mut self,
block_id: &BlockId,
header: WrappedHeader
) -> Result<(), GraphError>

Mark a block as invalid

pub fn get_genesis_block_ids(&self) -> &Vec<BlockId>

get genesis block ids

pub fn compute_max_cliques(&self) -> Vec<PreHashSet<BlockId>>

Computes max cliques of compatible blocks

pub fn prune(&mut self) -> Result<PreHashMap<BlockId, ActiveBlock>>

prune and return final blocks, return discarded final blocks

pub fn get_block_wishlist(
&self
) -> Result<PreHashMap<BlockId, Option<WrappedHeader>>>

get the current block wish list, including the operations hash.

pub fn get_clique_count(&self) -> usize

get clique count

pub fn get_blockclique(&self) -> &PreHashSet<BlockId>

get the clique of higher fitness

pub fn get_blockclique_block_at_slot(&self, slot: &Slot) -> Option<BlockId>

get the blockclique (or final) block ID at a given slot, if any

pub fn get_latest_blockclique_block_at_slot(&self, slot: &Slot) -> BlockId

get the latest blockclique (or final) block ID that is the most recent, but still strictly older than slot, in the same thread as slot

pub fn get_all_final_blocks(&self) -> HashMap<BlockId, Slot>

Gets all stored final blocks, not only the still-useful ones This is used when initializing Execution from Consensus. Since the Execution bootstrap snapshot is older than the Consensus snapshot, we might need to signal older final blocks for Execution to catch up.

pub fn get_blocks_to_propagate(&mut self) -> PreHashMap<BlockId, Storage>

Get the block id’s to be propagated. Must be called by the consensus worker within block_db_changed.

pub fn get_attack_attempts(&mut self) -> Vec<BlockId>

Get the hashes of objects that were attack attempts. Must be called by the consensus worker within block_db_changed.

pub fn get_new_final_blocks(&mut self) -> PreHashSet<BlockId>

Get the ids of blocks that became final. Must be called by the consensus worker within block_db_changed.

pub fn get_new_stale_blocks(&mut self) -> PreHashMap<BlockId, (Address, Slot)>

Get the ids of blocks that became stale. Must be called by the consensus worker within block_db_changed.