Struct massa_graph::BlockGraphExport

pub struct BlockGraphExport {
    pub genesis_blocks: Vec<BlockId>,
    pub active_blocks: Map<BlockId, ExportCompiledBlock>,
    pub discarded_blocks: Map<BlockId, (DiscardReason, SignedHeader)>,
    pub best_parents: Vec<(BlockId, u64)>,
    pub latest_final_blocks_periods: Vec<(BlockId, u64)>,
    pub gi_head: Map<BlockId, Set<BlockId>>,
    pub max_cliques: Vec<Clique>,
}

Fields
genesis_blocks: Vec<BlockId>

Genesis blocks.

active_blocks: Map<BlockId, ExportCompiledBlock>

Map of active blocks, were blocks are in their exported version.

discarded_blocks: Map<BlockId, (DiscardReason, SignedHeader)>

Finite cache of discarded blocks, in exported version.

best_parents: Vec<(BlockId, u64)>

Best parents hashe in each thread.

latest_final_blocks_periods: Vec<(BlockId, u64)>

Latest final period and block hash in each thread.

gi_head: Map<BlockId, Set<BlockId>>

Head of the incompatibility graph.

max_cliques: Vec<Clique>

List of maximal cliques of compatible blocks.

Implementations

impl<'a> BlockGraphExport

pub fn extract_from(
block_graph: &'a BlockGraph,
slot_start: Option<Slot>,
slot_end: Option<Slot>
) -> Self

Conversion from blockgraph.