Crate massa_graph

Re-exports

pub use settings::LedgerConfig;

Modules

error

export_active_block

ledger

settings

Structs

BlockGraph

BlockGraphExport

BlockStateAccumulator

Aggregated changes made during a block’s execution

BootstrapableGraph

ExportCompiledBlock

The block version that can be exported. Note that the detailed list of operation is not exported

LedgerDataExport

Enums

DiscardReason

ExportBlockStatus

Block status in the graph that can be exported.

Status

Functions

create_genesis_block

Creates genesis block in given thread.