Struct massa_graph::BlockGraph

pub struct BlockGraph {
    pub gi_head: Map<BlockId, Set<BlockId>>,
    /* private fields */
}

Fields
gi_head: Map<BlockId, Set<BlockId>>

Incompatibility graph: maps a block id to the block ids it is incompatible with One entry per Active Block

Implementations

impl BlockGraph

pub async fn new(
cfg: GraphConfig,
init: Option<BootstrapableGraph>
) -> Result<Self>

Creates a new block_graph.

Argument

cfg : consensus configuration.

serialization_context: SerializationContext instance

pub fn export_bootstrap_graph(&self) -> Result<BootstrapableGraph>

pub fn block_state_try_apply_op(
&self,
state_accu: &mut BlockStateAccumulator,
header: &SignedHeader,
operation: &SignedOperation,
pos: &mut ProofOfStake
) -> Result<()>

Try to apply an operation in the context of the block

Arguments

state_accu: where the changes are accumulated while we go through the block

header: the header of the block we are inside

operation: the operation that we are trying to apply

pos: proof of stake engine (used for roll related operations)

pub fn block_state_sync_rolls(
&self,
accu: &mut BlockStateAccumulator,
header: &SignedHeader,
pos: &ProofOfStake,
involved_addrs: &Set<Address>
) -> Result<()>

loads missing block state rolls if available

Arguments

accu: accumulated block changes

header: block header

pos: proof of state engine

involved_addrs: involved addresses

pub fn block_state_try_apply(
&self,
accu: &mut BlockStateAccumulator,
header: &SignedHeader,
opt_ledger_changes: Option<LedgerChanges>,
opt_roll_updates: Option<RollUpdates>,
pos: &mut ProofOfStake
) -> Result<()>

try to apply ledger/roll changes to a block state accumulator if it fails, the state should remain undisturbed

pub fn block_state_accumulator_init(
&self,
header: &SignedHeader,
pos: &mut ProofOfStake
) -> Result<BlockStateAccumulator>

initializes a block state accumulator from a block header

pub fn get_latest_final_blocks_periods(&self) -> &Vec<(BlockId, u64)>

Gets latest final blocks (hash, period) for each thread.

pub fn get_best_parents(&self) -> &Vec<(BlockId, u64)>

Gets best parents.

pub fn get_cliques(&self) -> Vec<Clique>

Gets the list of cliques.

pub fn get_block_ids_by_creator(
&self,
address: &Address
) -> Map<BlockId, Status>

Returns the list of block IDs created by a given address, and their finality statuses

pub fn get_roll_data_at_parent(
&self,
block_id: BlockId,
addrs_opt: Option<&Set<Address>>,
pos: &ProofOfStake
) -> Result<(RollCounts, RollUpdates)>

for algo see pos.md if addrs_opt is Some(addrs), restrict to addrs. If None, return all addresses. returns (roll_counts, cycle_roll_updates)

pub fn get_ledger_data_export(
&self,
addresses: &Set<Address>
) -> Result<LedgerDataExport>

gets Ledger data export for given Addressees

pub fn get_operations_involving_address(
&self,
address: &Address
) -> Result<Map<OperationId, OperationSearchResult>>

pub fn get_active_block(&self, block_id: &BlockId) -> Option<&ActiveBlock>

Gets whole compiled block corresponding to given hash, if it is active.

Argument

block_id : block ID

pub fn get_export_block_status(
&self,
block_id: &BlockId
) -> Option<ExportBlockStatus>

pub fn get_operations(
&self,
operation_ids: &Set<OperationId>
) -> Map<OperationId, OperationSearchResult>

Retrieves operations from operation Ids

pub fn slot_tick(
&mut self,
pos: &mut ProofOfStake,
current_slot: Option<Slot>
) -> Result<()>

signal new slot

pub fn incoming_header(
&mut self,
block_id: BlockId,
header: SignedHeader,
pos: &mut ProofOfStake,
current_slot: Option<Slot>
) -> Result<()>

A new header has come !

Checks performed:

Ignore genesis blocks.

See process.

pub fn incoming_block(
&mut self,
block_id: BlockId,
block: Block,
operation_set: Map<OperationId, (usize, u64)>,
endorsement_ids: Map<EndorsementId, u32>,
pos: &mut ProofOfStake,
current_slot: Option<Slot>
) -> Result<()>

A new block has come

Checks performed:

Ignore genesis blocks.

See process.

pub fn get_genesis_block_ids(&self) -> &Vec<BlockId>

pub fn get_ledger_at_parents(
&self,
parents: &[BlockId],
query_addrs: &Set<Address>
) -> Result<LedgerSubset>

Compute ledger subset after given parents for given addresses

pub fn compute_max_cliques(&self) -> Vec<Set<BlockId>>

Computes max cliques of compatible blocks

pub fn prune(&mut self) -> Result<Map<BlockId, ActiveBlock>>

prune and return final blocks, return discarded final blocks

pub fn get_block_wishlist(&self) -> Result<Set<BlockId>>

get the current block wishlist

pub fn get_clique_count(&self) -> usize

pub fn get_blockclique(&self) -> Set<BlockId>

pub fn clone_all_final_blocks(&self) -> Map<BlockId, Block>

Clones all stored final blocks, not only the still-useful ones This is used when initializing Execution from Consensus. Since the Execution bootstrap snapshot is older than the Consensus snapshot, we might need to signal older final blocks for Execution to catch up.

pub fn get_blocks_to_propagate(
&mut self
) -> Map<BlockId, (Block, Set<OperationId>, Vec<EndorsementId>)>

Get the headers to be propagated. Must be called by the consensus worker within block_db_changed.

pub fn get_attack_attempts(&mut self) -> Vec<BlockId>

Get the hashes of objects that were attack attempts. Must be called by the consensus worker within block_db_changed.

pub fn get_new_final_blocks(&mut self) -> Set<BlockId>

Get the ids of blocks that became final. Must be called by the consensus worker within block_db_changed.

pub fn get_new_stale_blocks(&mut self) -> Map<BlockId, (PublicKey, Slot)>

Get the ids of blocks that became stale. Must be called by the consensus worker within block_db_changed.

pub fn get_endorsement_by_address(
&self,
address: Address
) -> Result<Map<EndorsementId, SignedEndorsement>>

pub fn get_endorsement_by_id(
&self,
endorsements: Set<EndorsementId>
) -> Result<Map<EndorsementId, EndorsementInfo>>